Remove "/local" and "/local-connect"
- Move all WebInstance-centric code to models.remote
- Merge WebInstance into RemoteInstance
