+++ app/models/remote/EventManagerProtocol.scala
@@ -1,0 +1,10 @@
+package models.remote
+
+trait EventManagerProtocol {
+  protected val JoinKey        = "join"
+  protected val ChatterKey     = "chatter"
+  protected val CommandKey     = "command"
+  protected val ResponseKey    = "response"
+  protected val QuitKey        = "quit"
+  protected val ViewUpdateKey  = "update"
+}
