+++ app/models/remote/WebContextProtocol.scala
@@ -1,0 +1,10 @@
+package models.remote
+
+trait WebContextProtocol {
+  protected val ObserverContext = "observer"
+  protected val TurtlesContext  = "turtles"
+  protected val LinksContext    = "links"
+  protected val PatchesContext  = "patches"
+  protected val RoomContext     = "room"
+  protected val NetLogoUsername = "netlogo"
+}
