--- app/views/local/client.scala.html
@@ -1,19 +1,0 @@
-@()(implicit request: RequestHeader)
-
-@views.html.core.component.netLogoView(true) {
-
-  <script type="text/javascript" src='@controllers.local.routes.Local.engine'></script>
-  @views.html.core.component.clientImports("chat-local-module")
-
-  <td width="3%"></td>
-  <td height="100%">
-    @views.html.core.component.chatPane(Html(""))
-  </td>
-
-  <script type="text/javascript" charset="utf-8">
-    exports.ChatGlobals.socketURL = '@controllers.local.routes.Local.handleSocketConnection().webSocketURL()';
-  </script>
-
-  @views.html.core.component.lateInitScripts()
-}
-
